requirements.txt
lxml~=4.9.1
requests~=2.28.2
pyasn1~=0.4.8
grpcio-tools==1.46.5
protobuf==3.20.3